Fix wifi-scanning crash in AP mode
The wifiInfo local variable was shadowing the global variable
which was used in all other functions, so it only got initialized
locally and all the other functions accessed the global NULL
pointer and crashed immediately. Looks like this code was
never tested.

Also add a small API quirk workaround for ESP32, it does
not have Hidden in the getNetworkInfo() function.