Split webserver_config page into smaller functions
This function is pretty difficult to scroll by, split into a
_body_get() and _body_post() function and reduce temporary
string() creation a bit. Also reserve memory upfront
to avoid reallocations, which are fragmenting the heap,
which makes it a lot more stable.

Last but no least this splitting reduces total stack usage.
In total this saves about 800 bytes of flash due to
removed String() usage.