Fix generation of JSON in writeConfig
Commit 1997bee505b202bff898c854f429fefa19672a57 broke saving
of configuration files, as now every datatype is saved as a
string, which in the case of bool is considered true even
if the string is "false". With that changing of configuration
was completely messed up.

Originally this code was using ArduinoJSON for serialization,
which was removed due to memory pressure concerns. This can
be fixed by using the DynamicJsonDocument from ArduinoJson 6.x.

This does not only fix the bug but also reduces code bloat
by 160 bytes. The memory pressure problem will be fixed in
a followup commit.