Reduce temporary string creation
Using a string as a return value has the disadvantage
that it creates a temporary String object on the caller
side that needs to be copied and destructed. We can
reduce 300 bytes of code plus heap churn from loop()
by calling the sensor functions to store the result
in a parameter passed by reference.