Size reduce the Luftdaten logo
Instead of duplicating this in every intl* file we can include
a common logo definition, as it seems to be the same everwhere
except for BG.

Also convert this to PNG as the SVG files are fairly bloated.
In order to reduce the sizes of PNGs, we need to switch to 4bit
color mode, which is just a minor color shift (imho).

This reduces BG firmware by 2.5kb flash and others by about 620 bytes.