Split SNTP handling into a pre and post connect step
In order to be able to use NTP servers given out by DHCP, we need
to call configTime() before starting to connect to Wifi so
that the DHCP process can add the preferred NTP servers.

Also, with Arduino Core 2.6.0 the SNTP callback no longer completes
inside setup(), we need to enter loop() until it gets actually
processed. So split the OTA steps into the SNTP callback and wait
with regular loop() work until SNTP is complete.