Rewrote SDS handling from scratch (Fixes #552)
The yield() based decoding of SDS sensor input wasn't very
reliable, so occassionally the buffer wasn't decoded at the
time of the start of the read, which meant we were skipping
measurements. Now we're blocking until this has happened,
plus we align the decoding in the main loop to minimize blocking.

Also this saves quite a bit of code size, almost 400 bytes of flash.