Update to Arduino Core 2.7.4 (Fixes #753)
This is a huge update, covering a much newer SoftwareSerial and
includes many fixes and optimisations. Unfortunately it deprecates
SPIFFS so we now have to live with compiler warnings until
we managed to find an upgrade path to littlefs.

Includes a fix for a WPA plain-text attack security issue so
we need to update, even though it increases FW quite a bit.