Remove favicon listing from header
The listing of a favion in the header causes at least Chrome to fetch
the icon immediately, while the page is still loading and validating
CSS and other assets. This leads to parallel execution of requests,
which is hanging as ESP8266 WebServer only processes one request at a
time. This can make the webui appear very unresponsive.

Switch to an implicit favicon.ico fetch, which the browser executes
last, after finishing to load the site content (e.g. after
First-Contentful-Paint).